remove auth feature from application, added clean up organization in Author and Article models
