+++ src/main/java/com/main/logicgate/article/ArticleRepository.java
@@ -1,0 +1,7 @@
+package com.main.logicgate.article;
+
+import com.main.logicgate.article.Article;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ArticleRepository extends JpaRepository<Article, Long> {
+}
