+++ src/main/java/com/main/logicgate/author/AuthorRepository.java
@@ -1,0 +1,5 @@
+package com.main.logicgate.author;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AuthorRepository extends JpaRepository<Author, Long> { }
