--- src/main/java/com/main/logicgate/dto/AuthenticationResponseDTO.java
@@ -1,15 +1,0 @@
-package com.main.logicgate.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationResponseDTO {
-    private String token;
-}
