--- src/main/java/com/main/logicgate/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.main.logicgate.repository;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.main.logicgate.model.UserModel;
-import org.springframework.data.domain.Example;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.security.core.userdetails.User;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<UserModel, Long> {
-    Optional<UserModel> findByEmail(String email);
-}
