+++ src/main/resources/application.properties
@@ -4,9 +4,9 @@ spring.datasource.password=${DB_PASSWORD}
 #property.value.key=${SECRET_KEY}
 spring.config.import=optional:secrets.yml
 
-spring.jpa.hibernate.ddl-auto=update
+#spring.jpa.hibernate.ddl-auto=update
 spring.java.show-sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.properties.hibernate.format_sql=true
-#spring.security.user.name=${SECURITY_USER_NAME}
-#spring.security.user.password=${SECURITY_PASSWORD}
+#spring.security.author.name=${SECURITY_USER_NAME}
+#spring.security.author.password=${SECURITY_PASSWORD}
